third_party/blink/web_tests/fast/css/invalidation/checked-pseudo-expected.txt

Use descendant invalidation set for :checked pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(checkbox, '').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(checkbox, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
